requirements.txt

python-dotenv==1.0.0
imaplib2==3.8
pdfminer.six==20221105
python-docx==0.8.11
pytesseract==0.3.10
Pillow==10.0.0
gspread==5.8.0
google-auth==2.17.3
googletrans==4.0.0-rc10.0-rc1
